Closed security holes when an authorized user could make requests on behalf of another

+++ Habr.Presentation/Controllers/CommentController.cs
@@ -1,4 +1,6 @@
 ﻿using Habr.BusinessLogic.Services.Interfaces;
+using Habr.Common.Exceptions;
+using Habr.Presentation.Extensions;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -9,6 +11,7 @@ namespace Habr.Presentation.Controllers
     public class CommentController : ControllerBase
     {
         private readonly ICommentService _commentService;
+
         public CommentController(ICommentService commentService)
         {
             _commentService = commentService;
@@ -28,8 +31,10 @@ namespace Habr.Presentation.Controllers
 
         [Authorize]
         [HttpPost]
-        public async Task<IActionResult> CreateComment([FromQuery] int userId, [FromQuery] string text, [FromQuery] int postId, [FromQuery] int? parentId)
+        public async Task<IActionResult> CreateComment([FromQuery] string text, [FromQuery] int postId, [FromQuery] int? parentId)
         {
+            int userId = HttpContext.User.Identity.GetAuthorizedUserId();
+
             if (parentId is null)
             {
                 await _commentService.CreateCommentAsync(userId, postId, text);
@@ -45,8 +50,16 @@ namespace Habr.Presentation.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteComment(int id)
         {
-            await _commentService.DeleteCommentAsync(id);
-            return StatusCode(204);
+            var comment = await _commentService.GetFullCommentByIdAsync(id);
+            if (comment is not null && comment.UserId == HttpContext.User.Identity.GetAuthorizedUserId())
+            {
+                await _commentService.DeleteCommentAsync(id);
+                return StatusCode(204);
+            }
+            else
+            {
+                throw new BadRequestException(Common.Resources.UserExceptionMessageResource.AccessError);
+            }
         }
     }
 }
