+++ Habr.Presentation/Controllers/PostController.cs
@@ -1,5 +1,7 @@
 ﻿using Habr.BusinessLogic.Services.Interfaces;
+using Habr.Common.Exceptions;
 using Habr.DataAccess;
+using Habr.Presentation.Extensions;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -11,6 +13,7 @@ namespace Habr.Presentation.Controllers
     {
         private readonly IPostService _postService;
         private readonly ICommentService _commentService;
+
         public PostController(IPostService postService, ICommentService commentService)
         {
             _postService = postService;
@@ -52,27 +55,43 @@ namespace Habr.Presentation.Controllers
 
         [Authorize]
         [HttpPost]
-        public async Task<IActionResult> CreatePost([FromQuery] string title, [FromQuery] string text, [FromQuery] int userId, [FromQuery] bool isPublished)
+        public async Task<IActionResult> CreatePost([FromQuery] string title, [FromQuery] string text, [FromQuery] bool isPublished)
         {
 
-            await _postService.CreatePostAsync(title, text, userId, isPublished);
+            await _postService.CreatePostAsync(title, text, HttpContext.User.Identity.GetAuthorizedUserId(), isPublished);
             return StatusCode(201);
+
         }
 
         [Authorize]
         [HttpPut]
         public async Task<IActionResult> UpdatePost([FromBody] Post post)
         {
-            await _postService.UpdatePostAsync(post);
-            return StatusCode(204);
+            if (post is not null && post.UserId == HttpContext.User.Identity.GetAuthorizedUserId())
+            {
+                await _postService.UpdatePostAsync(post);
+                return StatusCode(204);
+            }
+            else
+            {
+                throw new BadRequestException(Common.Resources.UserExceptionMessageResource.AccessError);
+            }
         }
 
         [Authorize]
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeletePost(int id)
         {
-            await _postService.DeletePostAsync(id);
-            return StatusCode(204);
+            var post = await _postService.GetFullPostByIdAsync(id);
+            if (post is not null && post.UserId == HttpContext.User.Identity.GetAuthorizedUserId())
+            {
+                await _postService.DeletePostAsync(id);
+                return StatusCode(204);
+            }
+            else
+            {
+                throw new BadRequestException(Common.Resources.UserExceptionMessageResource.AccessError);
+            }
         }
     }
 }
